# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "selfevals",
    "description": "HTTP bridge for the selfevals evals framework.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/workspaces/{workspace_id}/datasets": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Datasets Index",
        "operationId": "datasets_index_api_workspaces__workspace_id__datasets_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DatasetStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by lifecycle status (draft/active/frozen/archived).",
              "title": "Status"
            },
            "description": "Filter by lifecycle status (draft/active/frozen/archived)."
          },
          {
            "name": "dataset_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by dataset type (e.g. capability, golden).",
              "title": "Dataset Type"
            },
            "description": "Filter by dataset type (e.g. capability, golden)."
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetListPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Datasets Create",
        "operationId": "datasets_create_api_workspaces__workspace_id__datasets_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatasetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/datasets/{dataset_id}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Datasets Show",
        "operationId": "datasets_show_api_workspaces__workspace_id__datasets__dataset_id__get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/datasets/upload": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Datasets Upload",
        "operationId": "datasets_upload_api_workspaces__workspace_id__datasets_upload_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_datasets_upload_api_workspaces__workspace_id__datasets_upload_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/datasets/{dataset_id}/freeze": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Datasets Freeze",
        "operationId": "datasets_freeze_api_workspaces__workspace_id__datasets__dataset_id__freeze_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/datasets/{dataset_id}/cases": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Datasets Append Case",
        "operationId": "datasets_append_case_api_workspaces__workspace_id__datasets__dataset_id__cases_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppendDatasetCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppendDatasetCaseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/datasets/{dataset_id}/baseline": {
      "get": {
        "tags": [
          "baseline"
        ],
        "summary": "Dataset Baseline Show",
        "operationId": "dataset_baseline_show_api_workspaces__workspace_id__datasets__dataset_id__baseline_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "baseline"
        ],
        "summary": "Dataset Baseline Set",
        "operationId": "dataset_baseline_set_api_workspaces__workspace_id__datasets__dataset_id__baseline_put",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBaselineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaselineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspace_id}/datasets/{dataset_id}/regression-check": {
      "post": {
        "tags": [
          "baseline"
        ],
        "summary": "Dataset Regression Check",
        "operationId": "dataset_regression_check_api_workspaces__workspace_id__datasets__dataset_id__regression_check_post",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dataset Id"
            }
          },
          {
            "name": "X-SelfEvals-User",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`.",
              "title": "X-Selfevals-User"
            },
            "description": "Caller identity. In `local`/`header` mode: a plain user id. In `token` mode: a signed token from `selfevals.api.tokens.issue_token`."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegressionCheckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegressionResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_datasets_upload_api_workspaces__workspace_id__datasets_upload_post": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Dataset name."
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "A .jsonl file, one case per line."
          },
          "dataset_type": {
            "type": "string",
            "title": "Dataset Type",
            "default": "capability"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "file"
        ],
        "title": "Body_datasets_upload_api_workspaces__workspace_id__datasets_upload_post"
      },
...
      "CreateDatasetRequest": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dataset_type": {
            "type": "string",
            "title": "Dataset Type",
            "default": "capability"
          },
          "cases": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cases"
          },
          "cases_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cases Path"
          },
          "split_allocation": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split Allocation"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateDatasetRequest",
        "description": "Create a dataset over HTTP (JSON body). Provide exactly one case source:\n\n* `cases` \u2014 the case dicts inline in the request body.\n* `cases_path` \u2014 a path to a JSONL file already on the server's disk.\n\n(To upload a `.jsonl` file directly, use the multipart `\u2026/datasets/upload`\nendpoint instead.) `dataset_type` defaults to `capability`; an optional\n`split_allocation` overrides the default 0.7/0.2/0.1 portfolio."
      },
...
      "DatasetListPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasetSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "DatasetListPage",
        "description": "Paginated `GET /workspaces/{ws}/datasets`."
      },
...
      "DatasetStatus": {
        "type": "string",
        "enum": [
          "draft",
          "frozen",
          "active",
          "archived"
        ],
        "title": "DatasetStatus",
        "description": "Lifecycle of a Dataset manifest."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SetBaselineRequest": {
        "properties": {
          "iteration_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iteration Id"
          }
        },
        "type": "object",
        "title": "SetBaselineRequest",
        "description": "Re-anchor a dataset's baseline. `iteration_id` omitted = use the best\ncompleted iteration on the dataset (same default as the CLI)."
      },
...
    }
  }
}